+++ CHANGELOG.rst
@@ -13,6 +13,7 @@ Changelog
   * Track.merge(), Backend.merge() and gpxdo merge have a new arg partial_tracks
   * new: gpxdo fix --add-minutes
   * new: gpxdo fix --simplify
+  * new: gpxdo fix --split
 
 1.3.2 release 2018-08-19
 ------------------------